+++ src/kz/greetgo/msoffice/xlsx/reader/model/FontData.java
@@ -1,7 +1,7 @@
 package kz.greetgo.msoffice.xlsx.reader.model;
 
 public class FontData {
-  public Integer size;
+  public Double size;
   public String name;
   public boolean bold = false;
 }
